requirements.txt
Flask
pymysql
google-cloud-pubsub
